Update peds script and add test set
