+++ test/peds_test/centos6/test/f4.out
@@ -1,0 +1,1 @@
+line3: we are in centos6
