+++ test/peds_test/conditions.txt
@@ -1,0 +1,2 @@
+centos6
+centos7
