--- test/peds_test/diff_file_centos6-7.txt
@@ -1,117 +1,0 @@
-monitor.txt 1
-release-notes/Structural_unproc.txt 0
-T1w/xfms/acpc.mat 1
-T1w/xfms/T1w_dc.nii.gz 1
-T1w/xfms/T2w_reg_dc.nii.gz 1
-T1w/T1w1_gdc.nii.gz 0
-T1w/T1w.nii.gz 0
-T1w/ACPCAlignment/roi2full.mat 0
-T1w/ACPCAlignment/robustroi.nii.gz 0
-T1w/ACPCAlignment/full2roi.mat 0
-T1w/ACPCAlignment/roi2std.mat 1
-T1w/ACPCAlignment/acpc_final.nii.gz 1
-T1w/ACPCAlignment/full2std.mat 1
-T1w/BrainExtraction_FNIRTbased/roughlin.mat 1
-T1w/T1w_acpc.nii.gz 1
-T1w/BrainExtraction_FNIRTbased/T1w_acpc_to_MNI_roughlin.nii.gz 1
-T1w/BrainExtraction_FNIRTbased/NonlinearReg.txt 1
-T1w/BrainExtraction_FNIRTbased/NonlinearReg.nii.gz 1
-T1w/BrainExtraction_FNIRTbased/str2standard.nii.gz 1
-T1w/BrainExtraction_FNIRTbased/NonlinearRegJacobians.nii.gz 1
-T1w/BrainExtraction_FNIRTbased/NonlinearIntensities.nii.gz.txt 1
-T1w/BrainExtraction_FNIRTbased/IntensityModulatedT1.nii.gz 1
-T1w/BrainExtraction_FNIRTbased/NonlinearIntensities.nii.gz 1
-T1w/BrainExtraction_FNIRTbased/T1w_acpc_to_MNI_nonlin.nii.gz 1
-T1w/BrainExtraction_FNIRTbased/standard2str.nii.gz 1
-T1w/T1w_acpc_brain_mask.nii.gz 1
-T1w/T1w_acpc_brain.nii.gz 1
-T1w/T1w_acpc_dc.nii.gz 1
-T1w/T1w_acpc_dc_brain.nii.gz 1
-T1w/T2w_acpc_dc.nii.gz 1
-T1w/BiasFieldCorrection_sqrtT1wXT1w/T1wmulT2w.nii.gz 1
-T1w/BiasFieldCorrection_sqrtT1wXT1w/T1wmulT2w_brain.nii.gz 1
-T1w/BiasFieldCorrection_sqrtT1wXT1w/T1wmulT2w_brain_norm.nii.gz 1
-T1w/BiasFieldCorrection_sqrtT1wXT1w/SmoothNorm_s5.nii.gz 1
-T1w/BiasFieldCorrection_sqrtT1wXT1w/T1wmulT2w_brain_norm_s5.nii.gz 1
-T1w/BiasFieldCorrection_sqrtT1wXT1w/T1wmulT2w_brain_norm_modulate.nii.gz 1
-T1w/BiasFieldCorrection_sqrtT1wXT1w/T1wmulT2w_brain_norm_modulate_mask.nii.gz 1
-T1w/BiasFieldCorrection_sqrtT1wXT1w/bias_raw.nii.gz 1
-T1w/BiasField_acpc_dc.nii.gz 1
-T1w/T1w_acpc_dc_restore_brain.nii.gz 1
-T1w/T1w_acpc_dc_restore.nii.gz 1
-T1w/T2w_acpc_dc_restore_brain.nii.gz 1
-T2w/xfms/acpc.mat 1
-T1w/T2w_acpc_dc_restore.nii.gz 1
-T2w/T2w1_gdc.nii.gz 0
-T2w/T2w.nii.gz 0
-T2w/ACPCAlignment/roi2full.mat 0
-T2w/ACPCAlignment/robustroi.nii.gz 0
-T2w/ACPCAlignment/full2roi.mat 0
-T2w/ACPCAlignment/roi2std.mat 1
-T2w/ACPCAlignment/full2std.mat 1
-T2w/ACPCAlignment/acpc_final.nii.gz 1
-T2w/T2w_acpc.nii.gz 1
-T2w/BrainExtraction_FNIRTbased/roughlin.mat 1
-T2w/BrainExtraction_FNIRTbased/T2w_acpc_to_MNI_roughlin.nii.gz 1
-T2w/BrainExtraction_FNIRTbased/NonlinearReg.txt 1
-T2w/BrainExtraction_FNIRTbased/NonlinearReg.nii.gz 1
-T2w/BrainExtraction_FNIRTbased/str2standard.nii.gz 1
-T2w/BrainExtraction_FNIRTbased/NonlinearRegJacobians.nii.gz 1
-T2w/BrainExtraction_FNIRTbased/IntensityModulatedT1.nii.gz 1
-T2w/BrainExtraction_FNIRTbased/NonlinearIntensities.nii.gz.txt 1
-T2w/BrainExtraction_FNIRTbased/NonlinearIntensities.nii.gz 1
-T2w/BrainExtraction_FNIRTbased/T2w_acpc_to_MNI_nonlin.nii.gz 1
-T2w/BrainExtraction_FNIRTbased/standard2str.nii.gz 1
-T2w/T2w_acpc_brain_mask.nii.gz 1
-T2w/T2w_acpc_brain.nii.gz 1
-T2w/T2wToT1wDistortionCorrectAndReg/FieldMap/Magnitude_brain_mask.nii.gz 0
-T2w/T2wToT1wDistortionCorrectAndReg/FieldMap/Magnitude_brain.nii.gz 0
-T2w/T2wToT1wDistortionCorrectAndReg/FieldMap/Magnitude.nii.gz 0
-T2w/T2wToT1wDistortionCorrectAndReg/FieldMap/Phase.nii.gz 0
-T2w/T2wToT1wDistortionCorrectAndReg/FieldMap/FieldMap.nii.gz 0
-T2w/T2wToT1wDistortionCorrectAndReg/Magnitude.nii.gz 0
-T2w/T2wToT1wDistortionCorrectAndReg/FieldMap.nii.gz 0
-T2w/T2wToT1wDistortionCorrectAndReg/Magnitude_brain.nii.gz 0
-T2w/T2wToT1wDistortionCorrectAndReg/FieldMap_ShiftMapT1w.nii.gz 0
-T2w/T2wToT1wDistortionCorrectAndReg/FieldMap_WarpT1w.nii.gz 0
-T2w/T2wToT1wDistortionCorrectAndReg/Magnitude_warppedT1w.nii.gz 0
-T2w/T2wToT1wDistortionCorrectAndReg/Fieldmap2T1w_acpc.mat 1
-T2w/T2wToT1wDistortionCorrectAndReg/Magnitude_warppedT1w2T1w_acpc.nii.gz 1
-T2w/T2wToT1wDistortionCorrectAndReg/FieldMap2T1w_acpc.nii.gz 1
-T2w/T2wToT1wDistortionCorrectAndReg/FieldMap2T1w_acpc_ShiftMap.nii.gz 1
-T2w/T2wToT1wDistortionCorrectAndReg/FieldMap2T1w_acpc_Warp.nii.gz 1
-T2w/T2wToT1wDistortionCorrectAndReg/T1w_acpc.nii.gz 1
-T2w/T2wToT1wDistortionCorrectAndReg/T1w_acpc_brain.nii.gz 1
-T2w/T2wToT1wDistortionCorrectAndReg/FieldMap_ShiftMapT2w.nii.gz 0
-T2w/T2wToT1wDistortionCorrectAndReg/FieldMap_WarpT2w.nii.gz 0
-T2w/T2wToT1wDistortionCorrectAndReg/Magnitude_warppedT2w.nii.gz 0
-T2w/T2wToT1wDistortionCorrectAndReg/Fieldmap2T2w_acpc.mat 1
-T2w/T2wToT1wDistortionCorrectAndReg/Magnitude_warppedT2w2T2w_acpc.nii.gz 1
-T2w/T2wToT1wDistortionCorrectAndReg/FieldMap2T2w_acpc.nii.gz 1
-T2w/T2wToT1wDistortionCorrectAndReg/FieldMap2T2w_acpc_ShiftMap.nii.gz 1
-T2w/T2wToT1wDistortionCorrectAndReg/FieldMap2T2w_acpc_Warp.nii.gz 1
-T2w/T2wToT1wDistortionCorrectAndReg/T2w_acpc.nii.gz 1
-T2w/T2wToT1wDistortionCorrectAndReg/T2w_acpc_brain.nii.gz 1
-T2w/T2wToT1wDistortionCorrectAndReg/T2w2T1w/T2w_reg.mat 1
-T2w/T2wToT1wDistortionCorrectAndReg/T2w2T1w/T2w_reg_fast_wmedge.nii.gz 1
-T2w/T2wToT1wDistortionCorrectAndReg/T2w2T1w/T2w_reg_init.mat 1
-T2w/T2wToT1wDistortionCorrectAndReg/T2w2T1w/T2w_reg.nii.gz 1
-T2w/T2wToT1wDistortionCorrectAndReg/T2w2T1w/T2w_dc_reg.nii.gz 1
-T2w/T2wToT1wDistortionCorrectAndReg/T2w2T1w/sqrtT1wbyT2w.nii.gz 1
-MNINonLinear/xfms/acpc2MNILinear.mat 1
-MNINonLinear/xfms/T1w_acpc_dc_restore_brain_to_MNILinear.nii.gz 1
-MNINonLinear/xfms/NonlinearReg.nii.gz 1
-MNINonLinear/xfms/NonlinearReg.txt 1
-MNINonLinear/xfms/acpc_dc2standard.nii.gz 1
-MNINonLinear/xfms/NonlinearRegJacobians.nii.gz 1
-MNINonLinear/xfms/NonlinearIntensities.nii.gz.txt 1
-MNINonLinear/xfms/IntensityModulatedT1.nii.gz 1
-MNINonLinear/xfms/NonlinearIntensities.nii.gz 1
-MNINonLinear/xfms/2mmReg.nii.gz 1
-MNINonLinear/xfms/standard2acpc_dc.nii.gz 1
-MNINonLinear/T1w.nii.gz 1
-MNINonLinear/T1w_restore.nii.gz 1
-MNINonLinear/T1w_restore_brain.nii.gz 1
-MNINonLinear/T2w.nii.gz 1
-MNINonLinear/T2w_restore.nii.gz 1
-MNINonLinear/T2w_restore_brain.nii.gz 1
