+++ test/peds_test/error_matrix.txt
@@ -1,0 +1,4 @@
+f1.out 1
+f2.out 0
+f3.out 1
+f4.out 1
