+++ test/peds_test/toremove.txt
@@ -1,0 +1,1 @@
+nothing
